feat: move session code to component
